fix: LeakMemoryDumper.h/GCRankBonusInfo.cpp x64 관련 오류 수정
- LeakMemoryDumper.h: placement operator new의 첫 매개변수가 unsigned int
  였는데 x64에서는 size_t(64비트)와 달라서 'operator new'의 첫 매개변수는
  size_t여야 한다는 오류(C2821)가 났음. size_t로 수정하고, DWORD 트래킹
  함수로 넘기는 포인터 캐스트도 uintptr_t를 경유하도록 정리.
- GCRankBonusInfo.cpp: write(m_RankBonusInfoList.size())가 size_t(64비트)를
  넘겨서 SocketOutputStream::write(uint)/write(ulong) 사이에서 모호했음
  (C2668). read()가 이 카운트를 BYTE 하나로 읽는 것과 맞춰 (BYTE) 캐스트로
  명시.

+++ Client/Packet/Gpackets/GCRankBonusInfo.cpp
@@ -59,7 +59,12 @@ void GCRankBonusInfo::write ( SocketOutputStream & oStream ) const
 	//--------------------------------------------------
 	// write pc type
 	//--------------------------------------------------
-	oStream.write( m_RankBonusInfoList.size() );
+	// read() reads this count back into a single BYTE (see below), so match
+	// that width here - m_RankBonusInfoList.size() is size_t (64-bit on
+	// x64), which is ambiguous between SocketOutputStream::write(uint)/
+	// write(ulong) (error C2668) and wasn't the wire format read() expects
+	// anyway.
+	oStream.write( (BYTE)m_RankBonusInfoList.size() );
 
 	for (std::list<DWORD>::const_iterator itr = m_RankBonusInfoList.begin(); itr!= m_RankBonusInfoList.end(); itr++) 
 	{
